Moved threading tests into test_iostream. Cleaned up some formatting. Deleted test_thread.{cpp,py}
