+++ include/pybind11/iostream.h
@@ -47,8 +47,8 @@ private:
             {
                 gil_scoped_acquire tmp;
 
-                // This subtraction cannot be negative, so dropping the sign
-                // this invokes python and therefore should be included in the
+                // This subtraction cannot be negative, so dropping the sign.
+                // This invokes python and therefore should be included in the
                 // GIL
                 str line(pbase(), static_cast<size_t>(pptr() - pbase()));
