+++ tests/test_iostream.cpp
@@ -11,6 +11,8 @@
 #include <pybind11/iostream.h>
 #include "pybind11_tests.h"
 #include <iostream>
+#include <thread>
+#include <atomic>
 
 
 void noisy_function(std::string msg, bool flush) {
@@ -24,6 +26,40 @@ void noisy_funct_dual(std::string msg, std::string emsg) {
     std::cout << msg;
     std::cerr << emsg;
 }
+
+// object to manage C++ thread
+// simply repeatedly write to std::cerr until stopped
+// redirect is called at some point to test the safety of scoped_estream_redirect
+struct TestThread {
+    TestThread() : t_{nullptr}, stop_{false} {
+        auto thread_f = [this] {
+            while( !this->stop_ ) {
+                std::cout << "x" << std::flush;
+                std::this_thread::sleep_for(std::chrono::microseconds(50));
+            } };
+        t_ = new std::thread(std::move(thread_f));
+    }
+
+    ~TestThread() {
+        delete t_;
+    }
+
+    void stop() { stop_ = 1; }
+
+    void join() {
+        py::gil_scoped_release gil_lock;
+        t_->join();
+    }
+
+    void sleep() {
+        py::gil_scoped_release gil_lock;
+        std::this_thread::sleep_for(std::chrono::milliseconds(50));
+    }
+
+    std::thread * t_;
+    std::atomic<bool> stop_;
+};
+
 
 TEST_SUBMODULE(iostream, m) {
 
@@ -70,4 +106,13 @@ TEST_SUBMODULE(iostream, m) {
         std::cout << msg << std::flush;
         std::cerr << emsg << std::flush;
     });
+
+    py::class_<TestThread>(m, "TestThread")
+        .def(py::init<>())
+
+        .def("stop", &TestThread::stop)
+
+        .def("join", &TestThread::join)
+
+        .def("sleep", &TestThread::sleep);
 }
